fixed issue with Sigma being always 0 fixed issue with L falling into [0, 100) range instead of [0,100]. 100 is a valid value for a white color.

+++ Foundation/Sources/Common/Foundation.Common/Media/ColorSpaces/LabColorSpace.cs
@@ -28,7 +28,7 @@ namespace SquaredInfinity.Foundation.Media.ColorSpaces
         /// <summary>
         /// 0.206896
         /// </summary>
-        public const double Sigma = 6 / 29;
+        public const double Sigma = (double) 6 / 29;
 
         /// <summary>
         /// 0.008856
